[Fix] Moved misplaced files
Moved some misplased files in the logger service.
Resolves #65.
