+++ server/services/logger-service/src/index.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import config from '../core/config';
-import makeExpressCallback from '../core/express-callback';
-import sharedQueue from '../core/interfaces/shared-queue';
-import {QueueMessage, Order} from '../core/@types/global';
+import config from '../../core/config';
+import makeExpressCallback from '../../core/express-callback';
+import sharedQueue from '../../core/interfaces/shared-queue';
+import {QueueMessage} from '../../core/@types/global';
 import {getOrderLogs, notFound} from './interfaces/controllers';
 import {storeLogMessage} from './usecases';
 
@@ -16,12 +16,10 @@ app.get(`${apiRoot}/logs/order`, makeExpressCallback(getOrderLogs));
 app.use(makeExpressCallback(notFound));
 
 setTimeout(() => sharedQueue.listen(
-    config.inboundPaymentServiceQueue,
-    async (message: QueueMessage<Order>) => {
-      storeLogMessage(message);
-    },
+    config.inboundLoggerServiceQueue,
+    async (message: QueueMessage<any>) => storeLogMessage(message),
 ), 10000);
 
-app.listen(3002, () => {
+app.listen(3003, () => {
   console.log('Logger service started on "/api/logs"...');
 });
